Initial commit. Can create a connection.
There is a connection type that will create a connection with the stomp
server when the new function is called. At the moment the connection is
create.

At the moment only the command of a frame can be parsed. This means when
the connection is created it only checks to see if the connection was
successful, it will not give any extra error information that would have
been sent back in the body of an ERROR frame.