Add parsing to parse_header.
Before this parse_header was stubbed out and returned ("", "") - a
header with an empty string as a key and an empty string as a value.
It now returns the real key and the value.

A lifetime had to be provided as the rust compiler needs to know that
the variable line stays around for as long as the keys and values do. I
think this is because split_str returns a vector of references, each to
a different part of line. Therefore line needs to stay valid whilst the
key and value are around. This is from my basic and probably faulty
knowledge of rust.