Add error checking to send_message.
send_message now checks what the server sends back. In order to not have
to complicate the code by checking for an ERROR frame which may never
come - and therefore having to deal with timeouts. I'm not sure how long
would be adequete or acceptable -  the function adds a receipt header so
that a frame will always be returned.