Make frame private in message.
The frame embedded in a message should not be accessible by users of the
library so this commit stops this behaviour.

Unfortunately the frame is still accessible by users of the library by
the get_frame method, however it can not be changed so this commit is
useful. Furthermore in practice users will not be able to access it as
frame is not exported.