Change message sending. Add option not to wait
Sometimes users of the library may not want to wait for a server to
confirm it has received a message, possibly because it takes too long or
it is not vital that it sent correctly.

The previous version tried to guarantee a server receieved a message
by asking for a RECEIPT frame.  This added a bug as a RECEIPT frame
may not be returned for every frame that asked for it because a RECEIPT
frame guarantees that the frames before it had been processed.

The new method is to allow a message to be sent in a fire and forget
way. Another method has been added which allows users to control how
long the connection waits for an ERROR frame.