Remove send_and_wait.
send_and_wait is removed from Connection as I think it is unlikely that
anyone would use it. Keeping it would make the code messy as it would
have to deal with getting a MESSAGE frame from one of the subscriptions
instead of an ERROR frame. For example in the last commit the message
was reported not to have sent correctly as a MESSAGE frame was received.