Add a task that reads from the stream.
Most stomp libraries allow you to pass a function that will be called
when you receive a message, for example stomp.py. They also will call
this function without you having to ask it to through a receive message.

To emulate this functionality a Task is spawned which reads from the
stream and then checks to see if it got an instruction to shut down. At
the moment if it receives a message it will only print it out.

Note that the example sleeps. If it did not do so than the message would
not be revceived as the task would be killed.