Make task call callback when it receives a message
When a user subscribes to a queue they need to pass a function. This
function is stored and will later be called if a message is received
from that queue. Ideally this would be a closure, however doing so means
adding lifetimes and it was difficult to get running. It may be worth
trying to do this later.

An Arc<Mutex> is used to store the subscriptions. This is because both
the users task and the reader task need to have access to the same data
(the Arc bit does that) and need to be able to write to it (the Mutex).

The receive method was removed because it is no longer needed and would
not work as expected.